update seller auth

+++ modules/sellers/App/Http/Controllers/Auth/CheckVerificationController.php
@@ -9,27 +9,22 @@ use Illuminate\Support\Facades\Password;
 use Illuminate\Contracts\Auth\PasswordBroker;
 use Modules\core\App\Models\VerificationCode;
 
-class CheckVerificationContoller extends Controller
+class CheckVerificationController extends Controller
 {
     public function __invoke(Request $request)
     {
         $username = $request->get('username');
         $code = $request->get('code');
-        $type = $request->get('type');
-        $seller = Seller::where('username', $username)->firstOrFail();
-        $verification = VerificationCode::where([
-            'tableable_type' => Seller::class,
-            'tableable_id' => $seller->id,
-            'code' => $code
-        ])->first();
-        if ($verification) {
-            $verification->delete();
-            if ($type == 'forget-password') {
-                $token = $this->broker()->sendResetLink(
-                    $request->only('username')
-                );
-                return ['status' => 'ok', 'token' => $token];
-            } else {
+        $seller = Seller::where('username', $username)
+            ->first();
+        if ($seller) {
+            $verification = VerificationCode::where([
+                'tableable_type' => Seller::class,
+                'tableable_id' => $seller->id,
+                'code' => $code
+            ])->first();
+            if ($verification) {
+                $verification->delete();
                 $seller->status = -2;
                 if (filter_var($seller->username, FILTER_VALIDATE_EMAIL)) {
                     $seller->email = $seller->username;
@@ -38,19 +33,17 @@ class CheckVerificationContoller extends Controller
                 }
                 $seller->update();
                 return ['status' => 'ok'];
+            } else {
+                return [
+                    'status' => 'error',
+                    'message' => 'کد وارد شده اشتباه است'
+                ];
             }
         } else {
             return [
                 'status' => 'error',
-                'message' => 'کد تایید وارد شده اشتباه می باشد',
+                'message' => 'فروشنده یافت نشد'
             ];
         }
-    }
-
-    protected function broker(): PasswordBroker
-    {
-        return Password::broker(
-            config('fortify.passwords')
-        );
     }
 }
